Refactor to clarify process
This refactor splits read into two function: read and forward. The former
returns a packet type that contains the information about what was read
and the information needed to forward a packet to the next hop.

The other major change is that the data consumed by each node is now the
pubkey of the next node it should forward the message to. As such ex1
consumes a fixed length of 33-bytes for the hopdata.  Ex2-4 consume
65-bytes (33 compress pubkey and 32 hmac). This is more akin to LN
which uses a fixed 65-bytes for legacy onion packets.